Refactor common.logging to include both syslog handlers
