+++ tubesync/common/logging/_filters.py
@@ -1,5 +1,4 @@
 import logging
-from logging.handlers import SysLogHandler
 
 
 class RemoveSpecificLogFilter(logging.Filter):
@@ -69,21 +68,7 @@ class RemoveSpecificLogFilter(logging.Filter):
         return False
 
 
-default_formatter = logging.Formatter(
-    '%(asctime)s [%(name)s/%(levelname)s] %(message)s'
-)
-default_handler = logging.StreamHandler()
-default_handler.setFormatter(default_formatter)
-default_handler.setLevel(logging.INFO)
-
-syslog_formatter = logging.Formatter(
-    '%(asctime)s %(name)s: %(message)s',
-    '%b %d %H:%M:%S',
-)
-syslog_handler = SysLogHandler(
-    address='/dev/log',
-    facility=SysLogHandler.LOG_LOCAL0,
-)
-syslog_handler.setFormatter(syslog_formatter)
-syslog_handler.setLevel(logging.DEBUG)
+__all__ = [
+    'RemoveSpecificLogFilter',
+]
