+++ tubesync/common/logging/syslog/__init__.py
@@ -1,0 +1,2 @@
+from . import hat as hat
+from . import std as std
