+++ tubesync/common/logging/syslog/hat/_logger.py
@@ -1,0 +1,7 @@
+import logging
+
+logger = lambda name=None: logging.getLogger(
+    __name__.rsplit('.', 1)[0] if name is None else name
+)
+
+__all__ = ['logger']
