+++ tubesync/common/logging/syslog/std/__init__.py
@@ -1,0 +1,3 @@
+from ._default import *  # noqa: F403
+from ._logger import logger as logger
+logger = logger(__name__)
